fix(38-01): invest 예측 RPC 2건 ROUND 인자에 ::numeric — 프로덕션 정의와 일치
파일↔프로덕션 drift 2건. PERCENTILE_CONT는 numeric 오버로드가 없어 ORDER BY 컬럼이
numeric이어도 double precision을 반환하므로, ROUND(<double precision>, 1) 호출이
check_function_bodies=on에서 CREATE 시점에 실패한다.

프로덕션 실측(pg_get_functiondef, invest_regional_prediction_summary):
  ROUND(br.median_change_pct::numeric, 1) AS median_change_pct
저장소 파일에는 ::numeric이 없었다 → 프로덕션 정의에 맞춰 추가.

대상:
- 20260601000001_invest_prediction_rpcs.sql:143
- 20260604000004_fix_prediction_ranking_future_only.sql:170

같은 함수의 다른 ROUND 호출(단일인자 ROUND(avg_*), ROUND((...)::numeric, 1))은
프로덕션과 이미 일치해 손대지 않았다. 프로덕션은 이미 이 상태이므로 db push 대상 아님.

+++ supabase/migrations/20260601000001_invest_prediction_rpcs.sql
@@ -140,7 +140,7 @@ AS $$
   SELECT
     br.sgg_code,
     br.complex_count,
-    ROUND(br.median_change_pct, 1) AS median_change_pct,
+    ROUND(br.median_change_pct::numeric, 1) AS median_change_pct,
     ROUND(ap.avg_near_price) AS avg_near_price,
     ROUND(ap.avg_far_price)  AS avg_far_price
   FROM by_region br
